+++ internal/app/storage/models/withdrawal.go
@@ -4,6 +4,6 @@ import "time"
 
 type Withdrawal struct {
 	Order        string    `json:"order"`
-	Sum          int       `json:"sum"`
+	Sum          float32   `json:"sum"`
 	Processed_at time.Time `json:"processed_at"`
 }
